Move Store methods. Add constants to users.go. Update newService signature
